requirements.txt
asgiref==3.7.2
boto3==1.28.24
botocore==1.31.24
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.6
Django==4.2.2
django-allauth==0.54.0
django-cors-headers==4.2.0
django-filter==23.2
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
filelock==3.12.0
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
jmespath==1.0.1
moviepy==1.0.3
numpy==1.25.2
oauthlib==3.2.2
Pillow==9.5.0
pipenv==2023.4.20
platformdirs==3.5.0
proglog==0.1.10
psutil==5.9.5
pycparser==2.21
PyJWT==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.6.3
urllib3==1.26.16
virtualenv==20.23.0
virtualenv-clone==0.5.7
